requirements.txt
ocr-nanonets-wrapper
os
json
cv2
numpy
pandas